Further cleanup of tr_gss and usage for tids handling
The trust router now builds, but the monitoring parser tests do not.

  * Eliminate extra layer of auth callback when using tr_gss.c, services
    using it now need only one auth callback
  * Document tr_gss.c's intended usage
  * Flesh out the MONS_INSTANCE structure
  * Fix a couple more pedantic data typing errors

+++ tr/tr_tid.c
@@ -698,14 +698,14 @@ int tr_tids_event_init(struct event_base *base,
   talloc_steal(tids, cookie);
 
   /* get a tids listener */
-  tids_ev->n_sock_fd = tids_get_listener(tids,
-                                         tr_tids_req_handler,
-                                         tr_tids_gss_handler,
-                                         cfg_mgr->active->internal->hostname,
-                                         cfg_mgr->active->internal->tids_port,
-                                         (void *)cookie,
-                                         tids_ev->sock_fd,
-                                         TR_MAX_SOCKETS);
+  tids_ev->n_sock_fd = (int)tids_get_listener(tids,
+                                              tr_tids_req_handler,
+                                              tr_tids_gss_handler,
+                                              cfg_mgr->active->internal->hostname,
+                                              cfg_mgr->active->internal->tids_port,
+                                              (void *)cookie,
+                                              tids_ev->sock_fd,
+                                              TR_MAX_SOCKETS);
   if (tids_ev->n_sock_fd==0) {
     tr_crit("Error opening TID server socket.");
     retval=1;
